+++ jscomp/syntax/src/react_jsx_common.ml
@@ -63,3 +63,15 @@ let removeArity binding =
     | _ -> expr
   in
   {binding with pvb_expr = removeArityRecord binding.pvb_expr}
+
+let async_component ~async expr =
+  if async then
+    let open Ast_helper in
+    Exp.apply
+      (Exp.ident
+         {
+           loc = Location.none;
+           txt = Ldot (Lident "JsxPPXReactSupport", "asyncComponent");
+         })
+      [(Nolabel, expr)]
+  else expr